+++ CHANGELOG.md
@@ -4,6 +4,34 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [4-r.1-beta.2] - 2023-01-26
+
+### Added
+
+* Add a description of type package to `README.md`.
+
+### Changed
+
+* Change Android SDK API level from 31 (Android 12) to 33 (Android 13).
+* Change the name and package of the `CubismRectangle` class to `type/csmRect` to match SDK for Native.
+* Move constants related to debugging from `CubismFramework` class to the newly created `CubismFrameworkConfig` class.
+* Change implementation to hide `CubismJsonString` from shallow layers. The following functions are affcted by this change.
+  * `getLayoutMap` function in `ICubismModelSetting` class
+  * `getLayoutMap` function in `CubismModelSettingJson` class
+  * `setupFromLayout` function in `CubismModelMatrix` class
+* Change the name and arguments of `createRenderer` function in `CubismUserModel`.
+  * The `RendererType` enumurator is abolished. Please generate a renderer you want to use by yourself and put it in the function as an argument.
+
+### Fixed
+
+* Fix JSON data parsing process to improve performance.
+* Fix a problem where `setClippingMaskBufferSize` in `CubismRendererAndroid` would be thrown a `NullPointerException` if there are no clipping masks in the model.
+
+### Removed
+
+* Remove dependencies not to be used.
+* Remove the unused method `getMotionMap` in `ICubismModelSetting` and `CubismModelSettingJson` class.
+
 ## [4-r.1-beta.1] - 2022-12-08
 
 ### Added
@@ -32,4 +60,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 * New released!
 
+[4-r.1-beta.2]: https://github.com/Live2D/CubismJavaFramework/compare/4-r.1-beta.1...4-r.1-beta.2
 [4-r.1-beta.1]: https://github.com/Live2D/CubismJavaFramework/compare/4-r.1-alpha.1...4-r.1-beta.1